Refactor shim functions into Arduino.h
